+++ src/lib.rs
@@ -67,7 +67,7 @@ mod util;
 pub use action::ActionType;
 pub use condition::*;
 pub use engine::{FilterEngine, FilterEngineBuilder};
-pub use r#enum::{FilterEnumItem, FilterEnumerator};
+pub use r#enum::{FilterEnumItem, FilterEnumerator, SubLayerEnumItem, SubLayerEnumerator};
 pub use filter::*;
 pub use layer::*;
 pub use provider::*;